Update README.md
Rockfall is an arcade-style space survival game where the player pilots a ship while avoiding falling asteroids, shoot asteroid that are moving close to the spacestation and surviving as long as possible. Built in Unity, the game features responsive controls, dynamic hazards, and a simple scoring system based on time and survivability.